Factoring some shared code
